trying new if function

+++ web/HoweCookbook/deleteRecipe.php
@@ -9,13 +9,15 @@ $stmt->execute();
 $recipeInfo = $stmt->fetch(PDO::FETCH_ASSOC);
 $recipeid = $recipeInfo['recipe_id'];
 
-$query1 = 'DELETE FROM recipes WHERE recipe_id =:recipeid';
-$stmt1 = $db->prepare($query1);
-$stmt1->bindValue(':recipe_id', $_GET['recipeId'], PDO::PARAM_INT);
-$stmt1->execute();
-header('Location: categoryRecipes.php');
-die();
-echo "just freakin work part 2";
+if ($_GET['recipeId']) {
+    echo "just freakin work part 2";
+    $query1 = 'DELETE FROM recipes WHERE recipe_id =:recipeid';
+    $stmt1 = $db->prepare($query1);
+    $stmt1->bindValue(':recipe_id', $_GET['recipeId'], PDO::PARAM_INT);
+    $stmt1->execute();
+    header('Location: categoryRecipes.php');
+    die();
+}
 
 ?>
 <!DOCTYPE html>
